app/requirements.txt
django>=4.0.0,<4.1.0
django-filter>=21.1.0,<21.2.0
djangorestframework>=3.13.1,<3.14.0
djangorestframework-simplejwt>=5.1.0,<5.2.0
psycopg2-binary>=2.9.3,<3.0.0
gunicorn>=20.1.0,<20.2.0